Move files to shared-styles
