+++ context/app/static/js/shared-styles/charts/AssayTypeBarChart/style.js
