+++ context/app/static/js/shared-styles/charts/AssayTypeBarChart/utils.js
